refactor: 기획 변경에 따라 페이지 값과 끝값여부를 반환하도록 수정
