+++ src/main/java/com/bookbla/americano/domain/book/service/BookProvider.java
@@ -1,10 +1,9 @@
 package com.bookbla.americano.domain.book.service;
 
-import com.bookbla.americano.domain.book.service.dto.BookInformationResponse;
-import java.util.List;
+import com.bookbla.americano.domain.book.service.dto.BookSearchResponses;
 
 public interface BookProvider {
 
-    List<BookInformationResponse> searchBooks(String text, int size, int page);
+    BookSearchResponses searchBooks(String text, int size, int page);
 
 }
